requirements-optional.txt
lxml
psycopg2
bs4
aria2p
googletrans
hachoir
emoji
Pillow
psutil
pySmartDL
regex
requests
speedtest-cli
sqlalchemy
asyncurban
oauth2client
google-api-python-client
httplib2==0.15.0
GitPython
coffeehouse
pymongo
heroku3
